[nrf noup] tests: ram_context_for_isr: Disable KMU by default
Disable the KMU by default for all the NRF54L devices for this
test. With the current configuration this test is incompatible
with the KMU because it changes the placement of the RAM in the
linker files and this collides with the KMU logic. But it also
does not need the KMU so there is no need to be there.

This is a noup because the configuration (and KMU support) is
only placed in ncs.

The KMU dependency to reserve the top RAM address is planned
to be done in dts (NCSDK-31980). Hopefuly when this is done
this noup can be removed.

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>